Get rid of the unnecessary loops
