+++ x-pack/plugins/apm/server/routes/traces/route.ts
@@ -8,7 +8,7 @@
 import * as t from 'io-ts';
 import { setupRequest } from '../../lib/helpers/setup_request';
 import { getTraceItems } from './get_trace_items';
-import { getTopTraces } from './get_top_traces';
+import { getTopTracesPrimaryStats } from './get_top_traces_primary_stats';
 import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
 import { environmentRt, kueryRt, rangeRt } from '../default_api_types';
 import { getSearchAggregatedTransactions } from '../../lib/helpers/transactions';
@@ -33,7 +33,7 @@ const tracesRoute = createApmServerRoute({
       end,
     });
 
-    return await getTopTraces({
+    return await getTopTracesPrimaryStats({
       environment,
       kuery,
       setup,
